requirements.txt
altgraph==0.17
asgiref==3.3.1
bcrypt==3.2.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
cx-Freeze==6.5.2
cx-logging==2.2
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dash-table-experiments==0.6.0
Django==3.0.10
django-mssql-backend==2.8.1
django-tinymce==3.2.0
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mssql-django==1.0b1
numpy==1.19.5
pandas==1.2.0
paramiko==2.7.2
pefile==2019.4.18
plotly==4.14.1
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyNaCl==1.4.0
pyodbc==4.0.30
pyserial==3.5
python-dateutil==2.8.1
pytz==2020.5
PyVISA==1.11.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.25.1
retrying==1.3.3
scapy==2.4.4
serial==0.0.97
six==1.15.0
sqlparse==0.4.1
tftpy==0.8.0
typing-extensions==3.7.4.3
urllib3==1.26.2
visa==1.0.0
visdcc==0.0.40
waitress==1.4.4
Werkzeug==1.0.1
zipp==3.4.0
djangorestframework==3.12.4